admin button is now hide if the user doesn't have admin role

+++ librarygateway/templates/static/lib/accueil.js
@@ -1,5 +1,11 @@
 window.onload = function() {
-	deleteAdmin()
+	let rawRoles  = sessionStorage.getItem('roles')
+	let roles = rawRoles.split(',')
+	if (roles.includes('administrator')){
+		showAdmin()
+	}else{
+		deleteAdmin()
+	}
 };
 
 function deleteAdmin(){
